More robust meta tag detection.
Meta tag detection used rexml. While this gives all the power of parsing
and working with XML, it is very strict on the input it accepts. On
bad markup it simply throw the towel and gave up on parsing.

As the detection of meta tags should not depend on bad (html) markup
much later in the file, the parsing is now ported to using Regular
Expression.