Put the options in proper order.
'global options' 'cover' 'toc options' 'page options'.

This does have a disadvantage: you cannot specify specific page options for toc or cover, and you can't put a cover in the middle of a pdf.

For example, if you'd like to add a page-number to your TOC with a footer, that's currently not possible, because footer is a page option. This could be fixed by parsing the content of the 'toc' parameter, and translate that into other options.