Pass response header 'wkhtmltopdf' to PDFKit, as options.
This lets the individual Rack endpoint control how the PDF is generated.
Currently, there's no good way to control settings on a
endpoint-by-endpoint basis.

Rack::Lint strongly suggests that the header values are all Strings, not
nested hashes, so they're passed as a JSON object. If the header is
missing, or the value is invalid JSON, it leaves it out.