Fix threadsafety issue.
This is a minimal fix. A more thorough fix might refactor the internals
of `#call` and its helpers to not rely on instance variables or mutable
state, but that would make it harder to tell whether threadsaftey, or
some other change, was the real bug.

This fix is borrowed from
https://github.com/ContinuityControl/pdfkit/commit/7c183c744218165dc39f6c6184a296aac6d4dc44

For more info, read:
* https://ieftimov.com/writing-rails-middleware#thread-safety
* http://stackoverflow.com/questions/23028226/rack-middleware-and-thread-safety

Fixes #358.