Remove middleware spec that examines mutated state.
The spec seems to be testing that the middleware's `@render_pdf` is
reset after each request. The spec now fails, because all of the
middleware's instance variables are, from the outside, irrelevant: they
only matter for the `dup` instance, which is GC'd right away.

I never like to delete a spec that starts to fail after a refactoring,
but I think it's OK here, because:
* the behavior the spec looks for is ensured by the threadsafety spec I
  added
* this spec exemplifies the problem with rack middleware: relying on
  mutable state