Handle case for possible $() execution
To make the tests pass, I had to add a way to check if shellcode could
have been used.  In doing so, $() is also something to worry about since
it will run shell code as well.  Unfortunately, I'm not as well versed
in all of the ways shell code could be executed here, so I attempted to
make it where you could add to the list to check against.