URI::parse unescaped source to re-encode
Instead of checking if it is a shell safe url, unencode it as it was
previously being tested, and then allow parse to attempt to re-encode it
for our use using it's to_s method.  This seems to pass all tests minus
two that are not expecting single quotes to be encoded, but in my tests,
urls placed in browser and in PDFKit.new.to_pdf seem to yield the same
results.

I did not want to change existing tests until getting feedback from
others before doing so as I do not want to cause unexpected results from
use, which this seems to do.